Set up conftest
